vendor in vendoer and some fixes

+++ routes/admin.php
@@ -55,22 +55,22 @@ Route::group(
                     Route::resourceWithDeleteSelected('roles', RoleController::class, [
                         'as' => 'admin'  // Prefix "admin." to all route names
                     ]);
-                    Route::get('activity_logs', [ActivityLogController::class, 'index'])->name('activity_logs.index');
-                    Route::delete('activity_logs/{id}', [ActivityLogController::class, 'destroy'])->name('activity_logs.destroy');
+                    Route::get('activity_logs', [ActivityLogController::class, 'index'])->name('admin.activity_logs.index');
+                    Route::delete('activity_logs/{id}', [ActivityLogController::class, 'destroy'])->name('admin.activity_logs.destroy');
                     #============================ User ====================================
 
                     #============================ User ====================================
 
                     #============================ vendors ====================================
 //                    Route::resourceWithDeleteSelected('vendors', VendorController::class);
-                    Route::get('vendors/index', [VendorController::class, 'index'])->name('vendors.index');
-                    Route::post('vendors/create', [VendorController::class, 'create'])->name('vendors.create');
-                    Route::put('vendors/update', [VendorController::class, 'update'])->name('vendors.update');
-                    Route::delete('vendors/{id}', [VendorController::class, 'destroy'])->name('vendors.destroy');
-                    Route::get('vendors/{id}/edit', [VendorController::class, 'edit'])->name('vendors.edit');
-//                    Route::get('vendors',[ VendorController::class,'index'])->name('vendors.index');
-                    Route::get('vendors/delete-selected', [VendorController::class, 'deleteSelected'])->name('vendors.deleteSelected');
-                    Route::post('vendors/update-column-selected', [VendorController::class, 'updateColumnSelected'])->name('vendors.updateColumnSelected');
+                    Route::get('vendors/index', [VendorController::class, 'index'])->name('admin.vendors.index');
+                    Route::get('vendors/create', [VendorController::class, 'create'])->name('admin.vendors.create');
+                    Route::post('vendors', [VendorController::class, 'store'])->name('admin.vendors.store');
+                    Route::put('vendors/update', [VendorController::class, 'update'])->name('admin.vendors.update');
+                    Route::delete('vendors/{id}', [VendorController::class, 'destroy'])->name('admin.vendors.destroy');
+                    Route::get('vendors/{id}/edit', [VendorController::class, 'edit'])->name('admin.vendors.edit');
+                    Route::get('vendors/delete-selected', [VendorController::class, 'deleteSelected'])->name('admin.vendors.deleteSelected');
+                    Route::post('vendors/update-column-selected', [VendorController::class, 'updateColumnSelected'])->name('admin.vendors.updateColumnSelected');
                     #============================ Admin ====================================
                     Route::resourceWithDeleteSelected('admins', AdminController::class);
                     #============================ countries ==================================
